refactored to move elasticsearch specific functionality to the elasticsearch backend and generic functionality out of the specific backend.
